Demo 1
The report displays the batches which need to be finished and shipped as soon as possible. As the report includes the raw materials’, the batches’, and the employees’ information, the management team of the brewery can monitor if the suppliers delivered the raw materials on time and know which employees should follow up and report the status of the process.